Project clean up

+++ Assets/Game/Scripts/Player.cs
@@ -1,16 +1,17 @@
 using System;
+using System.Numerics;
 using UnityEngine;
 
 namespace Etern0nety.Clicker
 {
     public class Player : MonoBehaviour
     {
-        public event Action<long> ScoreChanged;
+        public event Action<BigInteger> ScoreChanged;
 
-        [SerializeField] private long _score = 0;
+        [SerializeField] private BigInteger _score = 0;
         [SerializeField] private int _scoreForTap = 1;
 
-        public long Score => _score;
+        public BigInteger Score => _score;
         public int ScoreForTap => _scoreForTap;
 
         public void AddScore(int score)
